requirements.txt
customtkinter==5.2.2
mysql-connector-python==8.1.0
